+++ src/components/notificationMessage/notificationMessage.component.ts
@@ -1,10 +1,9 @@
 import {Component, EventEmitter, ChangeDetectionStrategy, ChangeDetectorRef, Inject, Optional, HostBinding} from '@angular/core';
-import {extend} from '@jscrpt/common';
+import {Notification, NotificationSeverity} from '@anglr/common';
 import {slideInOutTrigger} from '@anglr/animations';
+import {extend} from '@jscrpt/common';
 
 import {NotificationMessage, NotificationMessageCss, NotificationMessageOptions, NOTIFICATION_MESSAGE_OPTIONS} from '../../common/notifications.interface';
-import {NotificationType} from '../../common/notification.type';
-import {Notification} from '../../common/notification';
 
 /**
  * Default options for notification message component
@@ -13,6 +12,7 @@ import {Notification} from '../../common/notification';
 const defaultOptions: NotificationMessageOptions<NotificationMessageCss> =
 {
     clickToClose: true,
+    maxLength: 500,
     cssClasses:
     {
         clickable: 'clickable',
@@ -26,36 +26,25 @@ const defaultOptions: NotificationMessageOptions<NotificationMessageCss> =
  */
 @Component(
  {
-    selector: "notification",
-    styles: [`
-        .clickable
-        {
-            cursor: pointer;
-        }
-
-        :host
-        {
-            display: block;
-            overflow: hidden;
-        }
-    `],
+    selector: 'notification',
+    styleUrls: ['notificationMessage.component.css'],
     templateUrl: 'notificationMessage.component.html',
     animations: [slideInOutTrigger],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class NotificationMessageComponent implements NotificationMessage<NotificationMessageCss, NotificationMessageOptions<NotificationMessageCss>>
 {
-    //######################### private fields #########################
+    //######################### protected fields #########################
 
     /**
      * Item holding notification information
      */
-    private _item: Notification;
+    protected _item: Notification;
 
     /**
      * Represents notification options instance
      */
-    private _options: NotificationMessageOptions<NotificationMessageCss>;
+    protected _options: NotificationMessageOptions<NotificationMessageCss>;
 
     //######################### public properties - template bidings #########################
 
@@ -63,7 +52,7 @@ export class NotificationMessageComponent implements NotificationMessage<Notific
      * Object representing css class definition
      * @internal
      */
-    public classObj: {[key: string]: any} = {};
+    public classObj: {[key: string]: boolean} = {};
 
     //######################### public properties - host #########################
 
@@ -85,7 +74,7 @@ export class NotificationMessageComponent implements NotificationMessage<Notific
     }
     public set item(item: Notification)
     {
-        this.classObj[`${this._options.cssClasses.messageTypePrefix}${NotificationType[item.type]}`] = true;
+        this.classObj[`${this._options.cssClasses.messageTypePrefix}${NotificationSeverity[item.severity].toLowerCase()}`] = true;
         this._item = item;
     }
 
@@ -108,7 +97,7 @@ export class NotificationMessageComponent implements NotificationMessage<Notific
     public closed: EventEmitter<Notification> = new EventEmitter<Notification>();
 
     //######################### constructor #########################
-    constructor(private _changeDetector: ChangeDetectorRef,
+    constructor(protected _changeDetector: ChangeDetectorRef,
                 @Inject(NOTIFICATION_MESSAGE_OPTIONS) @Optional() options?: NotificationMessageOptions<NotificationMessageCss>)
     {
         this._options = extend(true, {}, defaultOptions, options);
@@ -120,7 +109,7 @@ export class NotificationMessageComponent implements NotificationMessage<Notific
      * Used for invoking 'closing' event
      * @internal
      */
-    public close()
+    public close(): void
     {
         if(this.options.clickToClose)
         {
@@ -138,12 +127,12 @@ export class NotificationMessageComponent implements NotificationMessage<Notific
         this._changeDetector.detectChanges();
     }
 
-    //######################### private methods #########################
+    //######################### protected methods #########################
 
     /**
      * Initialize component from provided options
      */
-    private _init()
+    protected _init(): void
     {
         this.classObj[this.options.cssClasses.clickable] = this.options.clickToClose;
     }