+++ src/components/notifications/notifications.component.ts
@@ -1,10 +1,10 @@
-import {Component, OnDestroy, Input, Optional, PLATFORM_ID, Inject, ChangeDetectionStrategy, ChangeDetectorRef} from '@angular/core';
+import {Component, OnDestroy, Input, Optional, PLATFORM_ID, Inject, ChangeDetectionStrategy, ChangeDetectorRef, OnInit} from '@angular/core';
 import {isPlatformBrowser} from '@angular/common';
+import {Notification} from '@anglr/common';
 import {extend} from '@jscrpt/common';
 import {Subscription} from 'rxjs';
 
-import {NotificationsOptions, NotificationMessageOptions, NOTIFICATIONS_OPTIONS} from '../../common/notifications.interface';
-import {Notification} from '../../common/notification';
+import {NotificationsOptions, NOTIFICATIONS_OPTIONS} from '../../common/notifications.interface';
 import {LocalNotificationsService} from '../../common/notifications.service';
 import {NotificationMessageComponent} from '../notificationMessage/notificationMessage.component';
 
@@ -12,14 +12,12 @@ import {NotificationMessageComponent} from '../notificationMessage/notificationM
  * Default options for notifications component
  * @internal
  */
-const defaultOptions: NotificationsOptions<any, NotificationMessageOptions<any>> =
+const defaultOptions: NotificationsOptions =
 {
     cssClasses:
     {
         rootDiv: 'notifications'
     },
-    maxLength: 500,
-    timeout: 10000,
     getNotificationMessageComponent: () => NotificationMessageComponent
 };
 
@@ -28,33 +26,23 @@ const defaultOptions: NotificationsOptions<any, NotificationMessageOptions<any>>
  */
 @Component(
 {
-    selector: "notifications",
+    selector: 'notifications',
     templateUrl: 'notifications.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class NotificationsComponent implements OnDestroy
+export class NotificationsComponent implements OnInit, OnDestroy
 {
-    //######################### private fields #########################
+    //######################### protected fields #########################
 
     /**
-     * Array of active timeouts
+     * Subscription all subscriptions created during initialization
      */
-    protected _timeouts: {[index: number]: any} = {};
-
-    /**
-     * Subscription for clearing event
-     */
-    protected _clearingSubscription: Subscription|null = null;
-
-    /**
-     * Subscription for notifying event
-     */
-    protected _notifyingSubscription: Subscription|null = null;
+    protected _initSubscriptions: Subscription = new Subscription();
 
     /**
      * Represents notification options instance
      */
-    protected _options: NotificationsOptions<any, NotificationMessageOptions<any>>;
+    protected _options: NotificationsOptions;
 
     //######################### public properties - template bindings #########################
 
@@ -62,11 +50,11 @@ export class NotificationsComponent implements OnDestroy
      * Options used for notification message
      */
     @Input()
-    public get options(): NotificationsOptions<any, NotificationMessageOptions<any>>
+    public get options(): NotificationsOptions
     {
         return this._options;
     }
-    public set options(options: NotificationsOptions<any, NotificationMessageOptions<any>>)
+    public set options(options: NotificationsOptions)
     {
         this._options = extend(true, this._options, options);
     }
@@ -77,88 +65,37 @@ export class NotificationsComponent implements OnDestroy
      * Array of displayed notifications - displayed set
      * @internal
      */
-    public notifications: Notification[] = [];
+    public notifications: readonly Notification[] = [];
 
     //######################### constructor #########################
-    constructor(service: LocalNotificationsService,
-                private _changeDetector: ChangeDetectorRef,
-                @Inject(PLATFORM_ID) platformId: Object,
-                @Inject(NOTIFICATIONS_OPTIONS) @Optional() options?: NotificationsOptions<any, NotificationMessageOptions<any>>)
+    constructor(protected _service: LocalNotificationsService,
+                protected _changeDetector: ChangeDetectorRef,
+                @Inject(PLATFORM_ID) protected _platformId: Object,
+                @Inject(NOTIFICATIONS_OPTIONS) @Optional() options?: NotificationsOptions)
     {
         this._options = extend(true, {}, defaultOptions, options);
+    }
 
-        if(!isPlatformBrowser(platformId))
+    //######################### public methods - implementation of OnInit #########################
+    
+    /**
+     * Initialize component
+     */
+    public ngOnInit(): void
+    {
+        if(!isPlatformBrowser(this._platformId))
         {
             return;
         }
 
-        //removing all displayed items
-        this._clearingSubscription = service.clearingMessages.subscribe(() =>
+        this.notifications = this._service.notifications;
+
+        this._initSubscriptions.add(this._service.notificationsChange.subscribe(() =>
         {
-            while(this.notifications.length > 0)
-            {
-                this.removeItem(this.notifications[0]);
-            }
+            this.notifications = this._service.notifications;
 
             this._changeDetector.detectChanges();
-        });
-
-        this._notifyingSubscription = service.notifying.subscribe((itm: Notification) =>
-        {
-            var id = 0;
-
-            if(this.notifications.length > 0)
-            {
-                id = this.notifications[this.notifications.length - 1].id + 1;
-            }
-
-            itm.id = id;
-
-            if(itm.message.length > this.options.maxLength)
-            {
-                itm.message = itm.message.substr(0, this.options.maxLength) + " ...";
-            }
-
-            if(this.options.timeout > 0)
-            {
-                this._timeouts[id] = setTimeout(() =>
-                {
-                    this.removeItem(itm);
-                    this._changeDetector.detectChanges();
-                }, this.options.timeout);
-            }
-
-            this.addItem(itm);
-
-            this._changeDetector.detectChanges();
-        });
-    }
-
-    //######################### public methods #########################
-
-    /**
-     * Adds notification item to list
-     * @param item - Item to be added
-     */
-    public addItem(item: Notification)
-    {
-        this.notifications.push(item);
-    }
-
-    /**
-     * Removes notification item from list
-     * @param item - Item to be removed
-     */
-    public removeItem(item: Notification)
-    {
-        var index = this.notifications.indexOf(item);
-
-        if (index > -1)
-        {
-            clearTimeout(this._timeouts[item.id]);
-            delete this._timeouts[item.id];
-            this.notifications.splice(index, 1);
-        }
+        }));
     }
 
     //######################### public methods - implementation of OnDestroy #########################
@@ -166,18 +103,22 @@ export class NotificationsComponent implements OnDestroy
     /**
      * Called when component is destroyed
      */
-    public ngOnDestroy()
+    public ngOnDestroy(): void
     {
-        if(this._clearingSubscription)
-        {
-            this._clearingSubscription.unsubscribe();
-            this._clearingSubscription = null;
-        }
+        this._initSubscriptions?.unsubscribe();
+        this._initSubscriptions = null;
+    }
 
-        if(this._notifyingSubscription)
-        {
-            this._notifyingSubscription.unsubscribe();
-            this._notifyingSubscription = null;
-        }
+    //######################### public methods #########################
+
+    /**
+     * Removes notification item from list
+     * @param item - Item to be removed
+     * 
+     * @internal
+     */
+    public removeItem(item: Notification): void
+    {
+        this._service.remove(item);
     }
 }