+++ src/directives/componentRenderer/messageRenderer.directive.ts
@@ -1,12 +1,12 @@
 import {ComponentRef, Directive, Input, OnChanges, SimpleChanges, ViewContainerRef, EventEmitter, Output, ComponentFactoryResolver, OnDestroy, Inject, Optional, Type} from '@angular/core';
+import {Notification} from '@anglr/common';
 import {nameof} from '@jscrpt/common';
 import {Subscription} from 'rxjs';
 
-import {NotificationMessage, NotificationMessageOptions, NotificationsOptions, NOTIFICATION_MESSAGE_SERVICE, NotificationMessageService} from '../../common/notifications.interface';
-import {Notification} from '../../common/notification';
+import {NotificationMessage, NotificationsOptions, NOTIFICATION_MESSAGE_SERVICE, NotificationMessageService} from '../../common/notifications.interface';
 
 /**
-* Creates dynamically instance of component by its metadata
+* Creates instance of component dynamically by its metadata
 */
 @Directive(
 {
@@ -15,17 +15,17 @@ import {Notification} from '../../common/notification';
 })
 export class MessageRendererDirective implements OnChanges, OnDestroy
 {
-    //######################### private fields #########################
+    //######################### protected fields #########################
 
     /**
      * Created component reference
      */
-    private _componentRef: ComponentRef<NotificationMessage<any, NotificationMessageOptions<any>>>|null = null;
+    protected _componentRef: ComponentRef<NotificationMessage>|null = null;
 
     /**
      * Subscription for listening on closed message
      */
-    private _closedSubscription: Subscription;
+    protected _closedSubscription: Subscription;
 
     //######################### public properties - inputs #########################
 
@@ -36,10 +36,10 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
     public notification: Notification;
 
     /**
-     * Options passed to notifications options
+     * Options passed to notifications component
      */
     @Input()
-    public notificationsOptions: NotificationsOptions<any, NotificationMessageOptions<any>>;
+    public notificationsOptions: NotificationsOptions;
 
     //######################### public properties - outputs #########################
 
@@ -49,12 +49,12 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
     @Output()
     public closed: EventEmitter<Notification> = new EventEmitter<Notification>();
 
-    //######################### private properties #########################
+    //######################### protected properties #########################
 
     /**
      * Instance of dynamically created component
      */
-    private get component(): NotificationMessage<any, NotificationMessageOptions<any>>|null
+    protected get component(): NotificationMessage|null
     {
         if(!this._componentRef)
         {
@@ -65,8 +65,8 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
     }
 
     //######################### constructor #########################
-    constructor(private _viewContainerRef: ViewContainerRef,
-                @Inject(NOTIFICATION_MESSAGE_SERVICE) @Optional() private _notificationMessageService?: NotificationMessageService<any, NotificationMessageOptions<any>>)
+    constructor(protected _viewContainerRef: ViewContainerRef,
+                @Inject(NOTIFICATION_MESSAGE_SERVICE) @Optional() protected _notificationMessageService?: NotificationMessageService)
     {
     }
 
@@ -75,22 +75,22 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
     /**
      * Called when input value changes
      */
-    public ngOnChanges(changes: SimpleChanges)
+    public ngOnChanges(changes: SimpleChanges): void
     {
         this._viewContainerRef.clear();
 
         if(nameof<MessageRendererDirective>('notification') in changes && changes[nameof<MessageRendererDirective>('notification')].currentValue)
         {
-            let injector = this._viewContainerRef.parentInjector;
-            let notificationMessageType: Type<NotificationMessage<any, NotificationMessageOptions<any>>>;
+            const injector = this._viewContainerRef.injector;
+            let notificationMessageType: Type<NotificationMessage>;
 
             if(this._notificationMessageService)
             {
-                notificationMessageType = this._notificationMessageService.getNotificationMessageComponent(this.notification.type);
+                notificationMessageType = this._notificationMessageService.getNotificationMessageComponent(this.notification.severity);
             }
             else
             {
-                notificationMessageType = this.notificationsOptions.getNotificationMessageComponent(this.notification.type);
+                notificationMessageType = this.notificationsOptions.getNotificationMessageComponent(this.notification.severity);
             }
 
             const componentFactoryResolver: ComponentFactoryResolver = injector.get(ComponentFactoryResolver);
@@ -101,7 +101,7 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
 
             this._closedSubscription = this.component.closed.subscribe(this._closedEmit);
             this.component.item = this.notification;
-            this.component.options = this.notificationsOptions.messageOptions || {};
+            this.component.options = this.notificationsOptions.messageOptions ?? {};
             this.component.invalidateVisuals();
         }
     }
@@ -111,22 +111,19 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
     /**
      * Called when component is destroyed
      */
-    public ngOnDestroy()
+    public ngOnDestroy(): void
     {
-        if(this._closedSubscription)
-        {
-            this._closedSubscription.unsubscribe();
-            this._closedSubscription = null;
-        }
+        this._closedSubscription?.unsubscribe();
+        this._closedSubscription = null;
     }
 
-    //######################### private methods #########################
+    //######################### protected methods #########################
 
     /**
      * Called when notification was closed by user
      * @param notification - Notification that was closed
      */
-    private _closedEmit = (notification: Notification) =>
+    protected _closedEmit = (notification: Notification): void =>
     {
         this.closed.emit(notification);
     }